requirements.txt
yfinance==0.2.36
pandas==2.1.4
redis==5.0.1
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
schedule==1.2.1
requests==2.31.0
openpyxl==3.1.2
httpx==0.26.0
pydantic==2.5.3
